bloglist

+++ controllers/blogs.js
@@ -2,59 +2,78 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
+const {userExtractor}  = require('../utils/middleware')
 
 
+/*
 const getTokenFrom = request => {
     const authorization = request.get('authorization')
-    if (authorization && authorization.startsWith('Bearer ')){
-        return authorization.replace('Bearer ', '')
+    if (authorization && authorization.toLowerCase().startsWith('bearer ')){
+        return authorization.replace('bearer ', '')
     }
     return null
 }
+*/
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
-    response.json(blogs)
+  const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
+  response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-    const body = request.body
+blogsRouter.post('/', userExtractor, async (request, response) => {
+  const body = request.body
 
-    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  //  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
 
+  //  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  /*
     if (!decodedToken.id) {
         return response.status(401).json({error: 'token invalid'})
     }
+*/
+  //    const user = await User.findById(decodedToken.id)
 
-    const user = await User.findById(decodedToken.id)
+  const user = request.user
 
-    const blog = new Blog( {
-        title: body.title,
-        author: body.author,
-        url: body.url,
-        likes: body.likes,
-        user: user._id
-    })
+  const blog = new Blog( {
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user._id
+  })
 
-    const savedBlog = await blog.save()
-    user.blogs = user.blogs.concat(savedBlog._id)
-    await user.save()
-    response.status(201).json(savedBlog)
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+  response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async(request, response) => {
+blogsRouter.delete('/:id', userExtractor, async(request, response) => {
+    
+  const blog = await Blog.findById(request.params.id)
+
+  const user = request.user
+
+
+  //  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  //   console.log(decodedToken)
+
+  if (blog.user.toString() === user.id.toString()){
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
+
+  }
 })
 
 blogsRouter.put('/:id', async(request, response) => {
-    const {title, author, url, likes} = request.body
+  const {title, author, url, likes} = request.body
 
-    const updatedBlog = await Blog.findByIdAndUpdate(
-        request.params.id,
-        {title, author, url, likes},
-        {new:true})
-        response.json(updatedBlog)
+  const updatedBlog = await Blog.findByIdAndUpdate(
+    request.params.id,
+    {title, author, url, likes},
+    {new:true})
+  response.json(updatedBlog)
 
 })
